requirements.txt
pyyaml
imageio
parameterized
ipywidgets
IPython
numpy==1.14.3
tensorflow
keras
scipy
mnist